# hatchery/openapi/model_status.go
/*
 * Hatchery API
 *
 * Workspace service for launching and interacting with containers.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Status struct {

	// Value:  * `Terminating` - The workspace is shutting down  * `Launching` - The workspace is starting up  * `Stopped` - The workspace is in a failed state and must be terminated  * `Running` - The workspace is running and ready to be used  * `Not Found` - The workspace could not be found 
	Status string `json:"status,omitempty"`

	// The condition of the pod
	Conditions []PodCondition `json:"conditions,omitempty"`

	// The state of all the containers
	ContainerStates []ContainerState `json:"containerStates,omitempty"`
}

// AssertStatusRequired checks if the required fields are not zero-ed
func AssertStatusRequired(obj Status) error {
	for _, el := range obj.Conditions {
		if err := AssertPodConditionRequired(el); err != nil {
			return err
		}
	}
	for _, el := range obj.ContainerStates {
		if err := AssertContainerStateRequired(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertRecurseStatusRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Status (e.g. [][]Status), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseStatusRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aStatus, ok := obj.(Status)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertStatusRequired(aStatus)
	})
}