# openapis/openapi.yaml
openapi: 3.0.1
info:
  title: Hatchery API
  description: Workspace service for launching and interacting with containers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://nci-crdc-demo.datacommons.io/lw-workspace/
tags:
- name: workspace
  description: Operations about workspaces
security:
  - bearerAuth: []
paths:
  /launch:
    post:
      tags:
      - workspace
      summary: LaunchAWorkspace
      operationId: launch
      parameters:
      - in: query
        name: id
        schema:
          type: string
        description: The ID of the workspace to launch from the /options list.
      - $ref: '#/components/parameters/User'
      - $ref: '#/components/parameters/Authorization'
      responses:
        200:
          description: successfully started launching
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /terminate:
    post:
      tags:
      - workspace
      summary: Terminate the actively running workspace
      operationId: terminate
      parameters:
      - $ref: '#/components/parameters/User'
      - $ref: '#/components/parameters/Authorization'
      - in: query
        name: id
        schema:
          type: string
        description: Workspace ID to be deleted
      responses:
        200:
          description: successfully started terminating
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /status:
    get:
      tags:
      - workspace
      summary: Get the current status of the workspace
      operationId: status
      parameters:
      - $ref: '#/components/parameters/User'
      - $ref: '#/components/parameters/Authorization'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /options:
    get:
      tags:
      - workspace
      summary: Get the available workspace options that can be launched
      operationId: options
      parameters:
      - $ref: '#/components/parameters/User'
      - $ref: '#/components/parameters/Authorization'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/Container'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /paymodels:
    get:
      tags:
      - workspace
      summary: Get the current user's pay model data
      operationId: paymodels
      parameters:
      - $ref: '#/components/parameters/User'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: '#/components/schemas/PayModel'
        401:
          $ref: '#/components/responses/UnauthorizedError'

components:
  parameters:
    User:
      in: header
      name: REMOTE_USER
      schema:
        type: string
    Authorization:
      in: header
      name: Authorization
      schema:
        type: string
  schemas:
    Status:
      type: object
      properties:
        status:
          type: string
          enum: [Launching, Running, Terminating, Stopped, Not Found]
          description: >
            Value:
             * `Terminating` - The workspace is shutting down
             * `Launching` - The workspace is starting up
             * `Stopped` - The workspace is in a failed state and must be terminated
             * `Running` - The workspace is running and ready to be used
             * `Not Found` - The workspace could not be found
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/PodCondition'
          description: The condition of the pod
        containerStates:
          type: array
          items:
            $ref: '#/components/schemas/ContainerState'
          description: The state of all the containers
    Container:
      type: object
      properties:
        name:
          type: string
          description: The display name for the container
        cpu-limit:
          type: string
          description: The CPU limit for the container
        memory-limit:
          type: string
          description: The memory limit for the container
        id:
          type: string
          description: The hash of the container, passed to /launch
    PodCondition:
      type: object
      properties:
        type:
          type: string
          enum: [PodScheduled, Initialized, ContainersReady, Ready]
          description: Name of this Pod condition
        status:
          type: string
          enum: ["True", "False", Unknown]
          description: Indicates whether that condition is applicable
    ContainerState:
      type: object
      properties:
        name:
          type: string
          description: Name of this container
        state:
          description: Details about the container's current condition
          $ref: "#/components/responses/ContainerStateDetail"
        ready:
          type: boolean
    ContainerStateDetail:
      type: object
      properties:
        running:
          nullable: true
          type: object
          properties:
            StartedAt:
              type: string
        waiting:
          nullable: true
          type: object
          properties:
            reason:
              type: string
            message:
              type: string
        terminated:
          nullable: true
          type: object
          properties:
            exitCode:
              type: int32
            signal:
              type: int32
            reason:
              type: string
            message:
              type: string
            startedAt:
              type: string
            finishedAt:
              type: string
            containerID:
              type: string
    PayModel:
      type: object
      properties:
        name:
          type: string
        user:
          type: string
        awsAccountId:
          type: string
        region:
          type: string
        ecs:
          type: string
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid